+++ weblate/trans/templatetags/translations.py
@@ -65,7 +65,6 @@ if TYPE_CHECKING:
 
     from django import forms
     from django.db.models import Model, QuerySet
-    from django.forms.boundfield import BoundField
     from django.template.context import Context
     from django.utils.safestring import SafeString
     from django_stubs_ext import StrOrPromise
@@ -1394,30 +1393,6 @@ def indicate_alerts(
 @register.filter(is_safe=True)
 def markdown(text: str) -> str:
     return format_html('<div class="markdown">{}</div>', render_markdown(text))
-
-
-@register.filter
-def choiceval(boundfield: BoundField) -> str:
-    """
-    Get literal value from a field's choices.
-
-    Empty value is returned if value is not selected or invalid.
-    """
-    value = boundfield.value()
-    if value is None:
-        return ""
-    if value is True:
-        return gettext("enabled")
-    if not hasattr(boundfield.field, "choices"):
-        return value
-    choices: dict[str, str] = {
-        str(choice): value for choice, value in boundfield.field.choices
-    }
-    if isinstance(value, list):
-        return format_html_join_comma(
-            "{}", list_to_tuples(choices.get(val, val) for val in value)
-        )
-    return choices.get(value, value)
 
 
 @register.filter